+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New features
 
 - Added `Wizaplace\SDK\Vendor\Attribute\AttributeService::getAttribute`
+- Added `Wizaplace\SDK\Commission\CommissionService::addCompanyCommission`
 - Added `Wizaplace\SDK\Commission\CommissionService::addMarketplaceCommission`
 - Added `Wizaplace\SDK\Commission\CommissionService::getMarketplaceCommission`
 - Added `Wizaplace\SDK\Commission\CommissionService::getCommission`